meeting_analyzer/requirements.txt
PyQt5>=5.15.0
openai-whisper>=20230314
spacy>=3.5.0
numpy>=1.24.0
sounddevice>=0.4.6

# After install, run: python -m spacy download en_core_web_sm
# sounddevice requires PortAudio:
#   Windows : included automatically via pip
#   Ubuntu  : sudo apt install libportaudio2
#   macOS   : brew install portaudio